Added Header Section and CSS animations Div.text-box.
