+++ public/src/app/common/components/home/home.component.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+  <link href="https://fonts.googleapis.com/css?family=Lato:100,300,400,700,900" rel="stylesheet">
+
+  <link rel="stylesheet" href="css/icon-font.css">
+  <link rel="stylesheet" href="css/styles.css">
+  <link rel="shortcut icon" type="image/png" href="img/favicon.png">
+
+  <title>Olango Paradise | Exciting tours for adventurous people.</title>
+</head>
+
+<body>
+  <header class="header">
+    <div class="logo-box">
+      <img src="img/logo-white.png" alt="Logo" class="logo">
+    </div>
+
+    <div class="text-box">
+      <h1 class="heading-primary">
+        <span class="heading-primary-main">Outdoors</span>
+        <span class="heading-primary-sub">is where life happens</span>
+      </h1>
+      <a href="#" class="btn btn-white btn-animated">Discover our tours</a>
+    </div>
+
+
+  </header>
+</body>
+
+</html>
