+++ public/src/app/common/components/home/home.component.scss
@@ -1,0 +1,22 @@
+/*
+COLORS:
+
+Light green: #7ed56f
+Medium green: #55c57a
+Dark green: #28b485
+
+*/
+
+* {
+    margin: 0px;
+    padding: 0px;
+    box-sizing: border-box;
+}
+
+body {
+    font-family: "Lato", sans-serif;
+    font-weight: 400;
+    font-size: 16PX;
+    line-height: 1.7;
+    color: #777;
+}
