+++ public/src/app/common/components/home/home.component.ts
@@ -1,0 +1,20 @@
+import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+import { HttpService } from '../../services/http.service';
+
+@Component({
+  selector: 'app-home',
+  templateUrl: './home.component.html',
+  styleUrls: ['./home.component.scss']
+})
+export class HomeComponent implements OnInit {
+
+  constructor(
+    private router: Router,
+    private http: HttpService,
+  ) { }
+
+  ngOnInit() {
+  }
+
+}
