--- public/src/app/http.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { HttpService } from './http.service';
-
-describe('HttpService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: HttpService = TestBed.get(HttpService);
-    expect(service).toBeTruthy();
-  });
-});
